+++ aiagents4pharma/talk2biomodels/configs/tools/__init__.py
@@ -3,3 +3,4 @@ Import all the modules in the package
 '''
 from . import ask_question
 from . import get_annotation
+from . import custom_plotter
