+++ aiagents4pharma/talk2biomodels/tools/parameter_scan.py
@@ -17,25 +17,11 @@ from langchain_core.messages import ToolMessage
 from langchain_core.tools.base import InjectedToolCallId
 from .load_biomodel import ModelData, load_biomodel
 from .load_arguments import TimeData, SpeciesInitialData
+from .utils import get_model_units
 
 # Initialize logger
 logging.basicConfig(level=logging.INFO)
 logger = logging.getLogger(__name__)
-
-def get_model_units(model_object):
-    """
-    Get the units of the model.
-
-    Args:
-        model_object: The model object.
-
-    Returns:
-        dict: The units of the model.
-    """
-    model_units = basico.model_info.get_model_units(model=model_object.copasi_model)
-    model_units_y = model_units['quantity_unit']
-    model_units_x = model_units['time_unit']
-    return {'y_axis_label': model_units_y, 'x_axis_label': model_units_x}
 
 @dataclass
 class ParameterScanData(BaseModel):
