+++ pyproject.toml
@@ -75,6 +75,7 @@ aiagents4pharma = [
   "talk2biomodels/configs/agents/t2b_agent/*",
   "talk2biomodels/configs/tools/ask_question/*",
   "talk2biomodels/configs/tools/get_annotation/*",
+  "talk2biomodels/configs/tools/custom_plotter/*",
   "talk2knowledgegraphs/configs/*",
   "talk2knowledgegraphs/configs/agents/t2kg_agent/*",
   "talk2knowledgegraphs/configs/app/frontend/*",
